Python/requirements.txt
flask==2.2.2
mysql-connector-python
Werkzeug==2.2.2
